+++ src/Aardvark.Base.Telemetry/Environment.cs
@@ -14,12 +14,16 @@ namespace Aardvark.Base
 
             public void SetMember(string name, IProbe probe)
             {
-                if (probe == null) throw new ArgumentNullException();
+                if (string.IsNullOrWhiteSpace(name)) throw new ArgumentNullException(nameof(name));
+                if (probe == null) throw new ArgumentNullException(nameof(probe));
                 m_probes[name] = new NamedProbe(name, probe);
             }
 
             public IProbe GetMember(string name)
-                => m_probes.TryGetValue(name, out NamedProbe x) ? x.Probe : null;
+            {
+                if (string.IsNullOrWhiteSpace(name)) throw new ArgumentNullException(nameof(name));
+                return m_probes.TryGetValue(name, out NamedProbe x) ? x.Probe : null;
+            }
 
             public override IEnumerable<string> GetDynamicMemberNames() => m_probes.Keys;
 
@@ -31,8 +35,10 @@ namespace Aardvark.Base
 
             public override bool TrySetMember(SetMemberBinder binder, object value)
             {
+                if (binder == null) throw new ArgumentNullException(nameof(binder));
+
                 var probe = value as IProbe;
-                if (probe == null) throw new ArgumentException();
+                if (probe == null) throw new ArgumentException("Value must implement Telemetry.IProbe.", nameof(value));
                 SetMember(binder.Name, probe);
                 return true;
             }